smoke-tests & buyer steps: do some file reorganization, renaming "marketplace" tests to "catalogue"
